requirements.txt
boto3
strands-agents
strands-agents-tools==0.2.1
aws-opentelemetry-distro==0.10.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit
flask-cors
plotly
python-dotenv